Fix square overflow in Vector::Normalised() (+1 squashed commits)
Squashed commits:

[7352e80] Fix overflow when dealing with squares of Fixed.
Treat ShapeData size as diameter instead of radius.

+++ LittleEngine/Source/Engine/Physics/CollisionManager.cpp
@@ -41,7 +41,7 @@ namespace Game {
 		for (size_t i = 0; i < colliders.size(); ++i) {
 			auto lhs = colliders[i].lock();
 			if (lhs != nullptr) {
-				for (int j = i + 1; j < colliders.size(); ++j) {
+				for (size_t j = i + 1; j < colliders.size(); ++j) {
 					auto rhs = colliders[j].lock();
 					if (rhs == nullptr) continue;
 					ProcessCollision(*lhs, *rhs);
